Make cargo test compile and pass, extract SlotTable into separate file
Fixed all test compilation issues by:
1. Identifying SlotTable-specific tests (those testing gaps, etc.) and ensuring
   they use SlotTable directly rather than SlotBackend
2. Updating helper function `apply_child_diff` to accept SlotBackend instead of
   SlotTable for general composition tests
3. Fixing setup/teardown functions to work with SlotBackend

Extracted SlotTable into separate file (slot_table.rs):
- SlotTable struct and all related types (Slot enum, SlotKind, GroupFrame)
- All SlotTable implementation methods
- SlotStorage trait implementation for SlotTable (baseline storage)
- Made mark_range_as_gaps public for test access

SlotTable is the baseline storage implementation, so no separate baseline
storage file is needed - it's already the reference implementation.

All 284 tests now pass successfully.

Key changes:
- crates/compose-core/src/slot_table.rs: New file with complete SlotTable impl
- crates/compose-core/src/lib.rs: Reduced by ~1600 lines, now imports SlotTable
- crates/compose-core/src/tests/lib_tests.rs: Fixed to work with extracted code

The codebase is now better organized with SlotTable in its own module.
